requirements.txt
Django>=5.2.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.1
django-filter>=23.5
yookassa>=3.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0

